+++ src/mass_driver/drivers/counter.py
@@ -3,9 +3,9 @@
 Simplest code we could implement that demonstrates PatchDriver capabilities.
 """
 
-from pathlib import Path
 
 from mass_driver.models.patchdriver import PatchDriver, PatchOutcome, PatchResult
+from mass_driver.models.repository import ClonedRepo
 
 
 class Counter(PatchDriver):
@@ -14,9 +14,9 @@ class Counter(PatchDriver):
     target_count: int
     counter_file: str
 
-    def run(self, repo: Path) -> PatchResult:
+    def run(self, repo: ClonedRepo) -> PatchResult:
         """Process the counter file"""
-        counter_filepath_abs = repo / self.counter_file
+        counter_filepath_abs = repo.cloned_path / self.counter_file
         if not counter_filepath_abs.is_file():
             return PatchResult(
                 outcome=PatchOutcome.PATCH_DOES_NOT_APPLY,